Implement a Source cache
This change implements a cache for Sources generated by this plugin
so that on subsequent builds (i.e. incremental builds) we can re-use
the same Source if the content hasn't changed.

This improves performance with Webpack's `futureEmitAssets` mode, where
there is a cache based on Source referential identity that determines
whether an asset should be re-emitted.

Fixes #7